Add timestamp_in_ns option
This reports timestamps in nanoseconds as a long long - which avoids the
loss of precision caused by the normal timestamps, where IEEE double
precision has rounding errors in the hundreds of ns.

My Cython skills are weak, so some of this could probably be written
better. It also assumes that the C compiler supports long long.